pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "project_analyzed"
version = "0.1.0"
description = "Un sistema di analisi dei requisiti basato su agenti AI"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.28.0",
    "crewai-tools>=0.40.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.8.1",
    "devtools>=0.12.2",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "greenlet>=3.1.1",
    "python-multipart>=0.0.20",
    "SQLAlchemy>=2.0.40",
    "aiosqlite>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.hatch.build.targets.wheel]
packages = ["project_analyzed"]